Execution_summary_chatbot/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
dataclasses-json==0.5.14
entrypoints==0.4
faiss-cpu==1.7.4
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gevent==24.2.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.48.2
httplib2==0.22.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==8.0.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.0.284
langchain-experimental==0.0.25
langsmith==0.0.92
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
networkx==3.3
numexpr==2.10.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pinecone-client==4.1.2
pinecone-plugin-interface==0.0.7
proto-plus==1.24.0
protobuf==3.19.6
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
Pympler==1.1
PyMySQL==1.1.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-file==2.1.0
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.31
streamlit==1.22.0
sympy==1.13.0
tbb==2021.13.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.4.0
tldextract==5.1.2
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
transformers==4.42.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
validators==0.33.0
watchdog==4.0.1
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.2
zope.event==5.0
zope.interface==6.4.post2